+++ src/test/java/one/util/streamex/MoreCollectorsTest.java
@@ -748,4 +748,14 @@ public class MoreCollectorsTest {
                     .dominators((a, b) -> a >= b));
         });
     }
+    
+    @Test
+    public void testMinMax() {
+        List<String> input = asList("abc", "a", "asdf", "gdasa", "gffsd", "sfgs", "b", "c", "dsgs");
+        checkCollector("minMax", Optional.of("agdasa"), input::stream, MoreCollectors.minMax(Comparator
+                .comparingInt(String::length), String::concat));
+        checkCollectorEmpty("minMax", Optional.empty(), MoreCollectors.minMax(Comparator.naturalOrder(), (min, max) -> {
+            throw new IllegalStateException("Should not be called");
+        }));
+    }
 }